Moved selection and entity into respective files
I further organized the code, moving the selection code into a Select class to encapsulate the logic better, and moving the entity/object/npc classes into their own files as well.

Hopefully this should make future reviews of the code much easier since we'll only need to look at small files for some changes. And the point of future changes can be more focused as well, if they happen to be isolated to a particular file.

I tried to ensure that 0 functional changes occurred with these organizational changes, and that the game still works exactly the same as before.
