+++ .gitignore
@@ -1,0 +1,5 @@
+
+__pycache__/shape.cpython-310.pyc
+__pycache__/entity.cpython-310.pyc
+__pycache__/selection.cpython-310.pyc
+.vscode/launch.json
